fix reference parsing when reference points to another reference
In cases where one reference points to another reference, for example,
when one cell has a formula definition that needs to be evaluated before
another cell can be evaluated, the RefValue would throw a circular
dependency even when it's not there.

i.e. if we have a table like this:
|   A  |   B   |
|:----:|:-----:|
| =1+1 | =A1+1 |

the reference of B1 would fail because A1 needs to be evaluated before
B1 can be evaluated. However, clearly, this case should not be identified
as a circular dependency, and instead, A1 should be evaluated first
followed by B1.

However, if we have a table like this:
|   A  |   B   |
|:----:|:-----:|
| =B1+1 | =A1+1 |

then we have a circular dependency because to calculate A1 we need to
calculate B1 but we cannot calculate B1 without A1, etc. and in those cases
we should still see a circular dependency being thrown.

There might be edges cases where this would blow up and I'd love to discuss :)